Remove auto-detected project icon feature
Retire src/scanner.ts in favor of a thin src/projectList.ts that reads
project rows, filters to on-disk paths, and projects each to
{ name, path, iconColor }. Drop the dead project-metadata code (language,
description, hasGit/hasClaude/lastModified) and the app-icon detection.

- Remove iconDataUrl/language/description/hasGit/hasClaude/lastModified
  from the Project interface; drop icon_data_url from ProjectRow and add()
- Add migration 011 dropping the icon_data_url column
- Remove the iconDataUrl <img> branch across HomeView, Sidebar, TitleBar,
  RecentTasksPanel, ResumeBanner, and IconColorSection; colored-initials
  badge is now the only project icon
- IconColorSection: make the whole row open the color picker, with hover
  feedback on the row and swatch
